requirements.txt
pandas
numpy
requests
python-dotenv
pyotp
openpyxl
urllib3
smartapi-python
websocket-client
logzero